cleanup + write pfs files to save directory
