+++ ultramodern/include/ultramodern/ultra64.h
@@ -74,6 +74,38 @@ typedef u64	OSTime;
 #define OS_EVENT_THREADSTATUS     13    /* CPU thread status: used by rmon */
 #define OS_EVENT_PRENMI           14    /* Pre NMI interrupt */
 
+/* Controller errors */
+
+#define CONT_NO_RESPONSE_ERROR          0x8
+#define CONT_OVERRUN_ERROR              0x4
+#define CONT_RANGE_ERROR               -1
+#define CONT_FRAME_ERROR                0x2
+#define CONT_COLLISION_ERROR            0x1
+
+/* Controller type */
+
+#define CONT_TYPE_NORMAL 0x0005
+#define CONT_TYPE_MOUSE  0x0002
+#define CONT_TYPE_VOICE  0x0100
+
+/* File System error number */
+
+#define PFS_ERR_NOPACK          1   /* no memory card is plugged or */
+#define PFS_ERR_NEW_PACK        2   /* ram pack has been changed to a different one */
+#define PFS_ERR_INCONSISTENT    3   /* need to run Pfschecker*/
+#define PFS_ERR_CONTRFAIL       CONT_OVERRUN_ERROR              
+#define PFS_ERR_INVALID         5   /* invalid parameter or file not exist*/
+#define PFS_ERR_BAD_DATA        6   /* the data read from pack are bad*/
+#define PFS_DATA_FULL           7   /* no free pages on ram pack*/
+#define PFS_DIR_FULL            8   /* no free directories on ram pack*/
+#define PFS_ERR_EXIST           9   /* file exists*/
+#define PFS_ERR_ID_FATAL        10  /* dead ram pack */
+#define PFS_ERR_DEVICE          11  /* wrong device type*/
+#define PFS_ERR_NO_GBCART       12  /* no gb cartridge (64GB-PAK) */
+#define PFS_ERR_NEW_GBCART      13  /* gb cartridge may be changed */
+
+/* File System size */
+
 #define PFS_INODE_SIZE_PER_PAGE 128
 #define PFS_FILE_NAME_LEN       16
 #define PFS_FILE_EXT_LEN        4
@@ -81,9 +113,19 @@ typedef u64	OSTime;
 #define PFS_ONE_PAGE            8   /* blocks */
 #define PFS_MAX_BANKS           62
 
+/* File System flag */
+
 #define PFS_READ                0
 #define PFS_WRITE               1
 #define PFS_CREATE              2
+
+/* File System status */
+
+#define PFS_INITIALIZED         0x1
+#define PFS_CORRUPTED           0x2
+#define PFS_ID_BROKEN           0x4
+#define PFS_MOTOR_INITIALIZED   0x8
+#define PFS_GBPAK_INITIALIZED   0x10
 
 #define	M_GFXTASK	1
 #define	M_AUDTASK	2
